Add files via upload
the code does not have the error in it anymore as I break the code before the index error.  I tried to break the code with an if statement but it did not work. as of now i dont know if i will make the animation an intended error or not